+++ lib/src/models/geolocation_model.dart
@@ -9,8 +9,26 @@ class Geolocation {
   });
 
   /// Creates an instance of [Geolocation] from a JSON map.
-  factory Geolocation.fromJson(Map<String, dynamic> json) => Geolocation(
-    lat: json["lat"],
-    long: json["long"],
-  );
+  factory Geolocation.fromJson(Map<String, dynamic> json) {
+    // Validation for lat
+    if (json['lat'] == null || json['lat'] is! String) {
+      throw const FormatException('The lat is invalid');
+    }
+
+    // Validation for long
+    if (json['long'] == null || json['long'] is! String) {
+      throw const FormatException('The long is invalid');
+    }
+
+    return Geolocation(
+      lat: json["lat"],
+      long: json["long"],
+    );
+  }
+
+  /// Returns a string representation of the [Geolocation] object.
+  @override
+  String toString() {
+    return 'Geolocation(lat: $lat, long: $long)';
+  }
 }