+++ CHANGELOG.md
@@ -1,3 +1,11 @@
-## 0.0.1
+## 0.0.1 - 2024-10-25
 
-* TODO: Describe initial release.
+- Initial implementation of the package that allows the Fake Store API to be consumed.
+- Simple example in the `example/` directory with three buttons to consume product, category and user endpoints.
+
+## 1.0.1 - 2024-10-31
+
+- Improved data validation for the `Product`, `Category` and `User` models to ensure that the values received are valid.
+- Improved API call through the `Repository` instead of the `Service`.
+- Improved state management in the `example/`.
+- Improved error handling to ensure that the application does not crash when an error occurs.